update treegrid finjs demo live config
