+++ live-editing/configs/TreeGridConfigGenerator.ts
@@ -1,6 +1,9 @@
 // tslint:disable:max-line-length
-import { IgxButtonModule, IgxDialogModule, IgxGridModule, IgxToggleModule, IgxTransactionService, IgxTreeGridModule } from "igniteui-angular";
+import { IgxButtonGroupModule, IgxButtonModule, IgxButtonModule, IgxDialogModule, IgxExcelExporterService, IgxGridModule, IgxIconModule, IgxRippleModule,
+    IgxSliderModule, IgxSwitchModule, IgxToggleModule, IgxToggleModule, IgxTransactionService, IgxTreeGridModule } from "igniteui-angular";
 import { TreeGridChilddatakeySampleComponent } from "../../src/app/tree-grid/tree-grid-childdatakey-sample/tree-grid-childdatakey-sample.component";
+import { TreeGridFinJSComponent } from "../../src/app/tree-grid/tree-grid-finjs/tree-grid-finjs-sample.component";
+import { TreeLocalDataService } from "../../src/app/tree-grid/tree-grid-finjs/treeLocalData.service";
 import { TreeGridPrimaryforeignkeySampleComponent } from "../../src/app/tree-grid/tree-grid-primaryforeignkey-sample/tree-grid-primaryforeignkey-sample.component";
 import { TreeGridRowEditSampleComponent } from "../../src/app/tree-grid/tree-grid-row-edit/tree-grid-row-editing-sample.component";
 import { AppModuleConfig } from "./core/AppModuleConfig";
@@ -45,6 +48,22 @@ export class TreeGridConfigGenerator implements IConfigGenerator {
             }),
             component: TreeGridRowEditSampleComponent,
             shortenComponentPathBy: "/tree-grid/"
+        }));
+
+        // TreeGrid FInJS demo component sample
+        configs.push(new Config({
+            additionalFiles: ["/src/app/tree-grid/tree-grid-finjs/treeLocalData.service.ts",
+                "/src/app/tree-grid/tree-grid-finjs/hierFinancialData.ts"],
+            appModuleConfig: new AppModuleConfig({
+                imports: [IgxGridModule, IgxButtonGroupModule, IgxIconModule, IgxSliderModule, IgxToggleModule,
+                    IgxButtonModule, IgxExcelExporterService, IgxSwitchModule, IgxRippleModule, TreeGridFinJSComponent,
+                    TreeLocalDataService],
+                ngDeclarations: [TreeGridFinJSComponent],
+                ngImports: [IgxGridModule, IgxButtonGroupModule, IgxIconModule, IgxSliderModule, IgxToggleModule,
+                    IgxButtonModule, IgxSwitchModule, IgxRippleModule, IgxGridModule.forRoot()],
+                ngProviders: [TreeLocalDataService, IgxExcelExporterService]
+            }),
+            component: TreeGridFinJSComponent
         }));
 
         return configs;
